Move login_for_access_token to auth router.py
