+++ backend/app/routers.py
@@ -1,9 +1,12 @@
 from fastapi import APIRouter
 
+from app.auth.router import router as auth_router
 from app.food_items.router import router as food_items_router
 from app.users.router import router as users_router
 
 api_router = APIRouter()
+
+api_router.include_router(auth_router, tags=["login"])
 
 api_router.include_router(
     food_items_router,
